requirements.txt
otree==3.3.11
psycopg2>=2.8.4
django-dotenv
requests
